apply Andy's feedback
