+++ tests/integration/schema/client_introspection/altair_graphiql_postman_2023.gql
