+++ tests/integration/schema/client_introspection/simple_test.go
@@ -30,29 +30,3 @@ func TestClientIntrospectionBasic(t *testing.T) {
 	}
 	testUtils.ExecuteTestCase(t, []string{}, test)
 }
-
-// TODO: This should pass without error, but we errors.
-func TestClientIntrospectionWithOneToManySchema(t *testing.T) {
-	test := testUtils.TestCase{
-		Actions: []any{
-			testUtils.SchemaUpdate{
-				Schema: `
-				type Book {
-					name: String
-					author: Author
-				}
-				type Author {
-					name: String
-					published: [Book]
-				}
-				`,
-			},
-			testUtils.ClientIntrospectionRequest{
-				Request:       clientIntrospectionQuery,
-				ExpectedError: "Unknown kind of type: ",
-				// ExpectedError: "InputFields are missing",
-			},
-		},
-	}
-	testUtils.ExecuteTestCase(t, []string{"Book", "Author"}, test)
-}
